requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-pinecone==0.0.1
pinecone-client==3.0.0